Create a idapro_interal utils file
